Login, registration, models, and db stuff
for landing.php
- index function checks if a user is logged and will show them the main
page or landing page(site login or registration) based on that
- registration function checks if a user has submitted registration form
and processes the information. Part of it isn't working yet.

--models: I added two models User_model and Profile_model. basically
these are used to access the database for users table and user_profile
table.
- I added two functions to the abstract model class Model.php in
codeigniter system/core/.
- The two functions are create(array of values, arrayof columns) and
exists(field, value). Ill update the readme for more about them.